Edit Quik 8.0.4.4 structs

+++ type_consistency/src/type_consistency.cpp
@@ -19,10 +19,10 @@ void test_getItem(qlua::api &q, std::function<void(const lua::entity<lua::type_p
   std::cout << "Table: ::qlua::table::" << table_name << std::endl;
   try {
     size_t n = q.getNumberOf<Table>();
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
       try {
         using entity = const lua::entity<lua::type_policy<Table>>&;
-        q.getItem<Table>(i, [&q, &f] (entity e) {
+        q.getItem<Table>(int(i), [&q, &f] (entity e) {
             f(e);
           });
       } catch (std::runtime_error e) {
